# api/openapi.yaml
openapi: 3.1.1
info:
  title: Cooking Assistant API
  version: 1.0.0
  description: API for recipe management and cooking assistance
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /recipes:
    get:
      tags: [Recipes]
      summary: List all recipes saved by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of user recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
        "401":
          description: Missing or invalid token

    post:
      tags: [Recipes]
      summary: Save a recipe to the current user's collection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Recipe saved successfully
        "400":
          description: Invalid recipe data
        "401":
          description: Missing or invalid token
  /recipes/{recipeId}:
    get:
      tags: [Recipes]
      summary: Get a specific recipe by ID
      security:
        - bearerAuth: []
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "400":
          description: Recipe ID is not a valid integer
        "401":
          description: Missing or invalid token
        "403":
          description: Recipe belongs to a different user
        "404":
          description: Recipe not found

  # -------------------------
  # AI
  # -------------------------
  /ai/recipes:
    post:
      tags: [AI]
      summary: Generate recipes using an LLM based on a prompt
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
      responses:
        "200":
          description: AI-generated recipes without IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecipeInput"
        "400":
          description: Invalid request body
        "401":
          description: Missing or invalid token
        "502":
          description: GenAI service unavailable or returned an unparseable response
  /ai/help:
    post:
      tags: [AI]
      summary: Ask the LLM a question about a recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequest"
      responses:
        "200":
          description: AI-generated help text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpResponse"
        "400":
          description: Invalid request body
        "401":
          description: Missing or invalid token
        "502":
          description: GenAI service unavailable or returned an unparseable response

# -------------------------
# SECURITY
# -------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------
  # SCHEMAS
  # -------------------------
  schemas:
...
    RecipeIngredient:
      type: object
      required: [quantity, unit, name]
      additionalProperties: false
      properties:
        quantity:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
          description: Unit of measurement (e.g. g, ml, cup, tbsp)
        name:
          type: string
          minLength: 1
    RecipeNutrients:
      type: object
      required: [calories, protein, fat, carbs]
      additionalProperties: false
      properties:
        calories:
          type: integer
          minimum: 0
        protein:
          type: integer
          minimum: 0
          description: Protein in grams
        fat:
          type: integer
          minimum: 0
          description: Fat in grams
        carbs:
          type: integer
          minimum: 0
          description: Carbohydrates in grams
    RecipeInput:
      type: object
      required: [title, ingredients, instructions, portions]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
          minItems: 1
        instructions:
          type: array
          items:
            type: string
            minLength: 1
          minItems: 1
        portions:
          type: integer
          minimum: 1
        nutrients:
          $ref: "#/components/schemas/RecipeNutrients"
    Recipe:
      type: object
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/RecipeInput"
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              minimum: 1
    RecipeRequest:
      type: object
      required: [prompt]
      additionalProperties: false
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
...
    HelpRequest:
      type: object
      required: [recipe, prompt]
      additionalProperties: false
      properties:
        recipe:
          $ref: "#/components/schemas/RecipeInput"
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
...
    HelpResponse:
      type: object
      required: [response]
      additionalProperties: false
      properties:
        response:
          type: string
          minLength: 1